+++ app/src/main/kotlin/com/wire/android/navigation/NavigationItem.kt
@@ -28,7 +28,7 @@ import com.wire.android.ui.home.HomeScreen
 import com.wire.android.ui.home.conversations.ConversationScreen
 import com.wire.android.ui.home.newconversation.NewConversationRouter
 import com.wire.android.ui.settings.SettingsScreen
-import com.wire.android.ui.userprofile.SelfUserProfileScreen
+import com.wire.android.ui.userprofile.self.SelfUserProfileScreen
 import com.wire.android.ui.userprofile.avatarpicker.AvatarPickerScreen
 import com.wire.kalium.logic.configuration.ServerConfig
 import com.wire.kalium.logic.data.conversation.ConversationId
