+++ app/src/main/kotlin/com/wire/android/ui/userprofile/self/model/OtherAccount.kt
@@ -1,0 +1,4 @@
+package com.wire.android.ui.userprofile.self.model
+
+
+data class OtherAccount(val id: String, val avatarUrl: String, val fullName: String, val teamName: String? = null)
